@@ -7,21 +7,20 @@ import android.view.View;
 import android.view.View.OnClickListener;
 import android.view.ViewGroup;
 import android.widget.Button;
+import android.widget.TextView;
 
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 import androidx.recyclerview.widget.RecyclerView.Adapter;
-import androidx.recyclerview.widget.RecyclerView.ViewHolder;
 
 import org.wordpress.android.R;
 import org.wordpress.android.WordPress;
 import org.wordpress.android.fluxc.store.AccountStore;
-import org.wordpress.android.ui.accounts.login.LoginHeaderViewHolder;
 import org.wordpress.android.ui.main.SitePickerAdapter;
-import org.wordpress.android.ui.main.SitePickerAdapter.HeaderHandler;
 import org.wordpress.android.ui.main.SitePickerAdapter.SiteList;
+import org.wordpress.android.ui.main.SitePickerAdapter.ViewHolderHandler;
 import org.wordpress.android.ui.media.MediaBrowserActivity;
 import org.wordpress.android.ui.posts.EditPostActivity;
 import org.wordpress.android.util.image.ImageManager;
@@ -33,7 +32,6 @@ public class ShareIntentReceiverFragment extends Fragment {
 
     private static final String ARG_SHARING_MEDIA = "ARG_SHARING_MEDIA";
     private static final String ARG_LAST_USED_BLOG_LOCAL_ID = "ARG_LAST_USED_BLOG_LOCAL_ID";
-    private static final String ARG_AFTER_LOGIN = "ARG_AFTER_LOGIN";
 
     @Inject AccountStore mAccountStore;
     @Inject ImageManager mImageManager;
@@ -45,17 +43,14 @@ public class ShareIntentReceiverFragment extends Fragment {
 
     private boolean mSharingMediaFile;
     private int mLastUsedBlogLocalId;
-    private boolean mAfterLogin;
     private RecyclerView mRecyclerView;
     private View mBottomButtonsContainer;
     private View mBottomButtonsShadow;
 
-    public static ShareIntentReceiverFragment newInstance(boolean sharingMediaFile, int lastUsedBlogLocalId,
-                                                          boolean afterLogin) {
+    public static ShareIntentReceiverFragment newInstance(boolean sharingMediaFile, int lastUsedBlogLocalId) {
         ShareIntentReceiverFragment fragment = new ShareIntentReceiverFragment();
         Bundle args = new Bundle();
         args.putBoolean(ARG_SHARING_MEDIA, sharingMediaFile);
-        args.putBoolean(ARG_AFTER_LOGIN, afterLogin);
         args.putInt(ARG_LAST_USED_BLOG_LOCAL_ID, lastUsedBlogLocalId);
         fragment.setArguments(args);
         return fragment;
@@ -81,7 +76,7 @@ public class ShareIntentReceiverFragment extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container,
                              @Nullable Bundle savedInstanceState) {
-        ViewGroup layout = (ViewGroup) inflater.inflate(R.layout.login_epilogue_screen, container, false);
+        ViewGroup layout = (ViewGroup) inflater.inflate(R.layout.share_intent_receiver_fragment, container, false);
         initButtonsContainer(layout);
         initShareActionPostButton(layout);
         initShareActionMediaButton(layout, mSharingMediaFile);
@@ -94,7 +89,6 @@ public class ShareIntentReceiverFragment extends Fragment {
         super.onCreate(savedInstanceState);
         ((WordPress) getActivity().getApplication()).component().inject(this);
         mSharingMediaFile = getArguments().getBoolean(ARG_SHARING_MEDIA);
-        mAfterLogin = getArguments().getBoolean(ARG_AFTER_LOGIN);
         mLastUsedBlogLocalId = getArguments().getInt(ARG_LAST_USED_BLOG_LOCAL_ID);
         loadSavedState(savedInstanceState);
     }
@@ -122,15 +116,12 @@ public class ShareIntentReceiverFragment extends Fragment {
     private void initShareActionPostButton(final ViewGroup layout) {
         mSharePostBtn = layout.findViewById(R.id.primary_button);
         addShareActionListener(mSharePostBtn, ShareAction.SHARE_TO_POST);
-        mSharePostBtn.setVisibility(View.VISIBLE);
-        mSharePostBtn.setText(R.string.share_action_post);
     }
 
     private void initShareActionMediaButton(final ViewGroup layout, boolean sharingMediaFile) {
         mShareMediaBtn = layout.findViewById(R.id.secondary_button);
         addShareActionListener(mShareMediaBtn, ShareAction.SHARE_TO_MEDIA_LIBRARY);
         mShareMediaBtn.setVisibility(sharingMediaFile ? View.VISIBLE : View.GONE);
-        mShareMediaBtn.setText(R.string.share_action_media);
     }
 
     private void addShareActionListener(final Button button, final ShareAction shareAction) {
@@ -184,28 +175,37 @@ public class ShareIntentReceiverFragment extends Fragment {
         return mAdapter;
     }
 
-    private HeaderHandler createHeaderHandler() {
-        return new HeaderHandler() {
+    private ViewHolderHandler<HeaderViewHolder> createHeaderHandler() {
+        return new ViewHolderHandler<HeaderViewHolder>() {
             @Override
-            public ViewHolder onCreateViewHolder(LayoutInflater layoutInflater, ViewGroup parent,
-                                                 boolean attachToRoot) {
-                return new LoginHeaderViewHolder(layoutInflater.inflate(R.layout.login_epilogue_header, parent, false));
+            public HeaderViewHolder onCreateViewHolder(LayoutInflater layoutInflater, ViewGroup parent,
+                                                       boolean attachToRoot) {
+                return new HeaderViewHolder(
+                        layoutInflater.inflate(R.layout.share_intent_receiver_header, parent, false));
             }
 
             @Override
-            public void onBindViewHolder(ViewHolder holder, SiteList sites) {
-                refreshAccountDetails((LoginHeaderViewHolder) holder, sites);
+            public void onBindViewHolder(HeaderViewHolder holder, SiteList sites) {
+                if (!isAdded()) {
+                    return;
+                }
+                holder.bindText(getString(
+                        sites.size() == 1 ? R.string.share_intent_adding_to : R.string.share_intent_pick_site));
             }
         };
     }
 
-    private void refreshAccountDetails(LoginHeaderViewHolder holder, SiteList sites) {
-        if (!isAdded()) {
-            return;
+    static class HeaderViewHolder extends RecyclerView.ViewHolder {
+        private final TextView mHeaderTextView;
+
+        HeaderViewHolder(View view) {
+            super(view);
+            mHeaderTextView = view.findViewById(R.id.login_epilogue_header_sites_subheader);
         }
-        holder.updateLoggedInAsHeading(getContext(), mImageManager, mAfterLogin, mAccountStore.getAccount());
-        holder.showSitesHeading(
-                getString(sites.size() == 1 ? R.string.share_intent_adding_to : R.string.share_intent_pick_site));
+
+        void bindText(String text) {
+            mHeaderTextView.setText(text);
+        }
     }
 
     enum ShareAction {
