@@ -16,11 +16,11 @@ import android.view.KeyEvent;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.view.ViewTreeObserver.OnGlobalLayoutListener;
 import android.widget.Button;
 import android.widget.EditText;
-import android.widget.ImageButton;
+import android.widget.FrameLayout;
 import android.widget.ImageView;
-import android.widget.RelativeLayout;
 import android.widget.TextView;
 
 import androidx.annotation.LayoutRes;
@@ -28,6 +28,8 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AlertDialog;
 import androidx.core.content.ContextCompat;
+import androidx.core.widget.NestedScrollView;
+import androidx.core.widget.NestedScrollView.OnScrollChangeListener;
 
 import com.google.android.material.dialog.MaterialAlertDialogBuilder;
 import com.yalantis.ucrop.UCrop;
@@ -55,6 +57,8 @@ import org.wordpress.android.ui.FullScreenDialogFragment;
 import org.wordpress.android.ui.FullScreenDialogFragment.OnConfirmListener;
 import org.wordpress.android.ui.FullScreenDialogFragment.OnDismissListener;
 import org.wordpress.android.ui.RequestCodes;
+import org.wordpress.android.ui.accounts.UnifiedLoginTracker;
+import org.wordpress.android.ui.accounts.UnifiedLoginTracker.Step;
 import org.wordpress.android.ui.media.MediaBrowserType;
 import org.wordpress.android.ui.photopicker.PhotoPickerActivity;
 import org.wordpress.android.ui.photopicker.PhotoPickerActivity.PhotoPickerMediaSource;
@@ -91,7 +95,7 @@ public class SignupEpilogueFragment extends LoginBaseFormFragment<SignupEpilogue
     private FullScreenDialogFragment mDialog;
     private SignupEpilogueListener mSignupEpilogueListener;
 
-    protected ImageButton mHeaderAvatarAdd;
+    protected ImageView mHeaderAvatarAdd;
     protected String mDisplayName;
     protected String mEmailAddress;
     protected String mPhotoUrl;
@@ -100,6 +104,8 @@ public class SignupEpilogueFragment extends LoginBaseFormFragment<SignupEpilogue
     protected ImageView mHeaderAvatar;
     protected WPTextView mHeaderDisplayName;
     protected WPTextView mHeaderEmailAddress;
+    protected View mBottomShadow;
+    protected NestedScrollView mScrollView;
     protected boolean mIsAvatarAdded;
     protected boolean mIsEmailSignup;
 
@@ -129,6 +135,7 @@ public class SignupEpilogueFragment extends LoginBaseFormFragment<SignupEpilogue
     @Inject protected Dispatcher mDispatcher;
     @Inject protected ImageManager mImageManager;
     @Inject protected AppPrefsWrapper mAppPrefsWrapper;
+    @Inject protected UnifiedLoginTracker mUnifiedLoginTracker;
 
     public static SignupEpilogueFragment newInstance(String displayName, String emailAddress,
                                                      String photoUrl, String username,
@@ -166,7 +173,7 @@ public class SignupEpilogueFragment extends LoginBaseFormFragment<SignupEpilogue
 
     @Override
     protected void setupContent(ViewGroup rootView) {
-        final RelativeLayout headerAvatarLayout = rootView.findViewById(R.id.signup_epilogue_header_avatar_layout);
+        final FrameLayout headerAvatarLayout = rootView.findViewById(R.id.login_epilogue_header_avatar_layout);
         headerAvatarLayout.setEnabled(mIsEmailSignup);
         headerAvatarLayout.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -185,12 +192,12 @@ public class SignupEpilogueFragment extends LoginBaseFormFragment<SignupEpilogue
             }
         });
         ViewUtilsKt.redirectContextClickToLongPressListener(headerAvatarLayout);
-        mHeaderAvatarAdd = rootView.findViewById(R.id.signup_epilogue_header_avatar_add);
+        mHeaderAvatarAdd = rootView.findViewById(R.id.login_epilogue_header_avatar_add);
         mHeaderAvatarAdd.setVisibility(mIsEmailSignup ? View.VISIBLE : View.GONE);
-        mHeaderAvatar = rootView.findViewById(R.id.signup_epilogue_header_avatar);
-        mHeaderDisplayName = rootView.findViewById(R.id.signup_epilogue_header_display);
+        mHeaderAvatar = rootView.findViewById(R.id.login_epilogue_header_avatar);
+        mHeaderDisplayName = rootView.findViewById(R.id.login_epilogue_header_title);
         mHeaderDisplayName.setText(mDisplayName);
-        mHeaderEmailAddress = rootView.findViewById(R.id.signup_epilogue_header_email);
+        mHeaderEmailAddress = rootView.findViewById(R.id.login_epilogue_header_subtitle);
         mHeaderEmailAddress.setText(mEmailAddress);
         WPLoginInputRow inputDisplayName = rootView.findViewById(R.id.signup_epilogue_input_display);
         mEditTextDisplayName = inputDisplayName.getEditText();
@@ -244,6 +251,27 @@ public class SignupEpilogueFragment extends LoginBaseFormFragment<SignupEpilogue
 
         // Set focus on static text field to avoid showing keyboard on start.
         mHeaderEmailAddress.requestFocus();
+
+        mBottomShadow = rootView.findViewById(R.id.bottom_shadow);
+        mScrollView = rootView.findViewById(R.id.scroll_view);
+        mScrollView.setOnScrollChangeListener(
+                (OnScrollChangeListener) (v, scrollX, scrollY, oldScrollX, oldScrollY) -> showBottomShadowIfNeeded());
+        // We must use onGlobalLayout here otherwise canScrollVertically will always return false
+        mScrollView.getViewTreeObserver().addOnGlobalLayoutListener(new OnGlobalLayoutListener() {
+            @Override public void onGlobalLayout() {
+                mScrollView.getViewTreeObserver().removeOnGlobalLayoutListener(this);
+                showBottomShadowIfNeeded();
+            }
+        });
+    }
+
+    private void showBottomShadowIfNeeded() {
+        if (mScrollView != null) {
+            final boolean canScrollDown = mScrollView.canScrollVertically(1);
+            if (mBottomShadow != null) {
+                mBottomShadow.setVisibility(canScrollDown ? View.VISIBLE : View.GONE);
+            }
+        }
     }
 
     @Override
@@ -278,6 +306,7 @@ public class SignupEpilogueFragment extends LoginBaseFormFragment<SignupEpilogue
             ReaderUpdateServiceStarter.startService(WordPress.getContext(),
                     EnumSet.of(ReaderUpdateLogic.UpdateTask.TAGS));
 
+            mUnifiedLoginTracker.track(Step.USERNAME_PASSWORD);
             if (mIsEmailSignup) {
                 AnalyticsTracker.track(AnalyticsTracker.Stat.SIGNUP_EMAIL_EPILOGUE_VIEWED);
 
