@@ -3,7 +3,6 @@ package org.wordpress.android.login;
 import android.app.ProgressDialog;
 import android.content.Context;
 import android.content.DialogInterface;
-import android.graphics.drawable.Drawable;
 import android.os.Bundle;
 import android.text.Html;
 import android.view.LayoutInflater;
@@ -23,13 +22,6 @@ import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.Toolbar;
 import androidx.fragment.app.Fragment;
 
-import com.bumptech.glide.Glide;
-import com.bumptech.glide.load.DataSource;
-import com.bumptech.glide.load.engine.GlideException;
-import com.bumptech.glide.request.RequestListener;
-import com.bumptech.glide.request.RequestOptions;
-import com.bumptech.glide.request.target.Target;
-
 import org.greenrobot.eventbus.Subscribe;
 import org.greenrobot.eventbus.ThreadMode;
 import org.wordpress.android.fluxc.Dispatcher;
@@ -39,8 +31,9 @@ import org.wordpress.android.fluxc.store.AccountStore.AuthEmailPayload;
 import org.wordpress.android.fluxc.store.AccountStore.AuthEmailPayloadScheme;
 import org.wordpress.android.fluxc.store.AccountStore.AuthEmailPayloadSource;
 import org.wordpress.android.fluxc.store.AccountStore.OnAuthEmailSent;
+import org.wordpress.android.login.util.AvatarHelper;
+import org.wordpress.android.login.util.AvatarHelper.AvatarRequestListener;
 import org.wordpress.android.util.AppLog;
-import org.wordpress.android.util.GravatarUtils;
 import org.wordpress.android.util.NetworkUtils;
 import org.wordpress.android.util.ToastUtils;
 
@@ -151,6 +144,9 @@ public class LoginMagicLinkRequestFragment extends Fragment {
         mAvatarProgressBar = view.findViewById(R.id.avatar_progress);
         ImageView avatarView = view.findViewById(R.id.gravatar);
 
+        TextView emailView = view.findViewById(R.id.email);
+        emailView.setText(mEmail);
+
         // Design changes added to the Woo Magic link sign-in
         if (mVerifyMagicLinkEmail) {
             View avatarContainerView = view.findViewById(R.id.avatar_container);
@@ -169,28 +165,11 @@ public class LoginMagicLinkRequestFragment extends Fragment {
 
             mRequestMagicLinkButton.setText(getString(R.string.send_verification_email));
         } else {
-            Glide.with(this)
-                 .load(GravatarUtils.gravatarFromEmail(mEmail,
-                         getContext().getResources().getDimensionPixelSize(R.dimen.avatar_sz_login)))
-                 .apply(RequestOptions.circleCropTransform())
-                 .apply(RequestOptions.placeholderOf(R.drawable.ic_gridicons_user_circle_100dp))
-                 .apply(RequestOptions.errorOf(R.drawable.ic_gridicons_user_circle_100dp))
-                 .listener(new RequestListener<Drawable>() {
-                     @Override
-                     public boolean onLoadFailed(@Nullable GlideException e, Object o, Target<Drawable> target,
-                                                 boolean b) {
-                         mAvatarProgressBar.setVisibility(View.GONE);
-                         return false;
-                     }
-
-                     @Override
-                     public boolean onResourceReady(Drawable drawable, Object o, Target<Drawable> target,
-                                                    DataSource dataSource, boolean b) {
-                         mAvatarProgressBar.setVisibility(View.GONE);
-                         return false;
-                     }
-                 })
-                 .into(avatarView);
+            AvatarHelper.loadAvatarFromEmail(this, mEmail, avatarView, new AvatarRequestListener() {
+                @Override public void onRequestFinished() {
+                    mAvatarProgressBar.setVisibility(View.GONE);
+                }
+            });
         }
 
         return view;
@@ -205,7 +184,7 @@ public class LoginMagicLinkRequestFragment extends Fragment {
 
         ActionBar actionBar = ((AppCompatActivity) getActivity()).getSupportActionBar();
         if (actionBar != null) {
-            actionBar.setDisplayShowTitleEnabled(false);
+            actionBar.setTitle(R.string.log_in);
             actionBar.setDisplayHomeAsUpEnabled(true);
         }
 
